requirements.txt
torchvision==0.18.1
ultralytics
supervision==0.2.0